Enhance mopro-ffi package: added README for documentation, updated check-mopro-download.sh to prevent redundant downloads, modified .gitignore to include additional framework files, and adjusted Android build configurations.

+++ check-mopro-download.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+if [ ! -d MoproFfiFramework.xcframework ]; then
 curl -o MoproFfiFramework.xcframework.zip https://ci-keys.zkmopro.org/MoproFfiFramework.xcframework.zip \
-&& unzip -o -q MoproFfiFramework.xcframework.zip \
-&& rm -f MoproFfiFramework.xcframework.zip
+  && unzip -o -q MoproFfiFramework.xcframework.zip \
+  && rm -f MoproFfiFramework.xcframework.zip
+fi
